% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointR.R
\name{check_tag}
\alias{check_tag}
\title{Add a new tag to a project stage}
\usage{
check_tag(stage, comment = "")
}
\arguments{
\item{stage}{Character. Required. The project stage.}

\item{comment}{Character. Optional. Description or comment for the tag.}
}
\value{
Invisibly returns a data.frame with the new tag.
}
\description{
Registers a comment tag for a project stage. A tag includes version, timestamp, and optional comment. Logged to `tags_log.xlsx`.
}
