% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointR.R
\name{check_tags}
\alias{check_tags}
\title{Show registered checkpoint tags}
\usage{
check_tags(stage = NULL, version = NULL)
}
\arguments{
\item{stage}{Optional. Character. Stage name. Use `"stages"` to show one tag per stage.}

\item{version}{Optional. Integer. Specific version to display from a stage.}
}
\value{
Invisibly returns NULL. Outputs tag data to the console.
}
\description{
Displays information from the tags log (`tags_log.xlsx`) including summaries, specific tags, or the latest tag.
}
