% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{checkpoint}
\alias{checkpoint}
\title{Evaluates an expression or loads a cached value}
\usage{
checkpoint(expr, ckpt.id, ..., force = FALSE, envir = parent.frame(),
  enclos = environment())
}
\arguments{
\item{expr}{any valid R expression}

\item{ckpt.id}{an identifier used for the checkpointed data}

\item{...}{any R object that is monitored for changed and is used to trigger re-evaluation}

\item{force}{force re-evaluation of the expression regardless of the state of `deps` or `expr`}

\item{envir}{the environment in which the expression is evaluated}

\item{enclos}{the environment in which the expression is substituted}
}
\value{
the result of the expression
}
\description{
Both the change to the environment and the expression's return value is stored.
The expression is re-evaluated if the expression or value of the dependent variables change.
}
