% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointR.R
\docType{package}
\name{checkpointR}
\alias{checkpointR-package}
\alias{checkpointR}
\title{checkpointR: Versioned Checkpoints for R Projects}
\description{
Tools for saving, tagging, loading and managing data object checkpoints across stages.
}
\keyword{internal}
