% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{copy.env}
\alias{copy.env}
\title{Helper function to copy the content of one environment to another}
\usage{
copy.env(source_env, destination_env, all.names = FALSE)
}
\arguments{
\item{source_env}{the environment from which data will be copied}

\item{destination_env}{the environment to which data will be copied}

\item{all.names}{whether or not to include hidden variables}
}
\description{
Helper function to copy the content of one environment to another
}
