% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{has.cache}
\alias{has.cache}
\title{Checks whether the checkpoint id exists}
\usage{
has.cache(ckpt.id, file)
}
\arguments{
\item{ckpt.id}{the checkpoint identifier}

\item{file}{the checkpoint file to check}
}
\value{
TRUE if the checkpoint id exists, FALSE if not
}
\description{
Checks whether the checkpoint id exists
}
