% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{restore.env}
\alias{restore.env}
\title{Restore an environment}
\usage{
restore.env(envir, ckpt.id, file)
}
\arguments{
\item{envir}{the environment in which to restore the variables}

\item{ckpt.id}{the environment id to load}

\item{file}{the file from which to load the environment}
}
\description{
Loads an environment by its id from the given file
}
