% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{store.env}
\alias{store.env}
\title{Copy an environment}
\usage{
store.env(envir, ckpt.id, file)
}
\arguments{
\item{envir}{the environment to copy}

\item{ckpt.id}{the identifier to associate the environment with}

\item{file}{the file that will contain the copy}
}
\description{
Copies an environment into a checkpoint file and associates it with an id.
Multiple environments can be stored in one file.
}
