% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointR.R
\name{check_load}
\alias{check_load}
\title{Load a saved checkpoint object from a given stage and version}
\usage{
check_load(
  stage,
  name = "procdata",
  version = NULL,
  folder = "4_checkpoint",
  envir = .GlobalEnv,
  quiet = FALSE
)
}
\arguments{
\item{stage}{Character. Stage from which to load the checkpoint.}

\item{name}{Character. Name of the object to load. Default is "procdata".}

\item{version}{Integer. Specific version to load. If NULL, loads the most recent.}

\item{folder}{Character. Folder where checkpoints are stored. Default is "4_checkpoint".}

\item{envir}{Environment. Where to assign the loaded object. Default is global environment.}

\item{quiet}{Logical. If TRUE, suppress console output. Default is FALSE.}
}
\value{
Invisible NULL. Assigns the object in the specified environment.
}
\description{
Load a saved checkpoint object from a given stage and version
}
