% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointR.R
\name{check_overview}
\alias{check_overview}
\title{Show an overview of available and loaded checkpoints}
\usage{
check_overview(stage = NULL, envir = .GlobalEnv)
}
\arguments{
\item{stage}{Optional character vector. If provided, filters by stage(s).}

\item{envir}{Environment. The environment where loaded objects are searched. Default is global.}
}
\value{
A list with two tibbles: available_versions and loaded_bases.
}
\description{
Show an overview of available and loaded checkpoints
}
